Dispatch TimesheetSubmitted event when submitted_at date is set.
Renamed TimesheetCompleted to TimesheetSubmitted to more accurately
represent its meaning. The SendTimesheetNotification listener sends
the Timesheet notification if the Timesheet's email_sent_at
attribute has not been set, and sets email_sent_at so that the
email won't be sent again.

This allows the TimesheetSubmitted event to be triggered without using
a state machine, which will allow the timesheet to be submitted and
emailed simply by setting the submitted_at date with a JSON:API call.